feat(navbar): Refactor Sw360 main NavBar
- Change double row layout to a traditional layout. This change sabe a
  lot of space on verical area
- Replace underline navigation to a traditional hover navigation
- Bring the header components over the navbar
